// src/main/resources/content/src/app/rmanUsers/rmanUsers.ts
export interface RmanUsers {
    attribute3: any;
    createdBy: any;
    attribute2: any;
    lastUpdatedBy: any;
    startDateActive: any;
    attribute1: any;
    rmanEmployeeId: any;
    creationDate: any;
    userName: any;
    lastName: any;
    lastUpdateDate: any;
    workPhone: any;
    firstName: any;
    attribute5: any;
    attribute4: any;
    cellPhone: any;
    emailAddress: any;
    userPassword: any;
    mgrId: any;
    fullName: any;
    homePhone: any;
    endDateActive: any;
    rmanUserId: any;
    enabledFlag: any;
}